Did a decent amount of cleanup and moved buttons so Zoe and I can integrate
